+++ src/Attributes/Validation/Same.php
@@ -3,13 +3,16 @@
 namespace Spatie\LaravelData\Attributes\Validation;
 
 use Attribute;
-use Spatie\LaravelData\Support\Validation\References\RouteParameterReference;
+use Spatie\LaravelData\Support\Validation\References\FieldReference;
 
 #[Attribute(Attribute::TARGET_PROPERTY)]
 class Same extends StringValidationAttribute
 {
-    public function __construct(protected string | RouteParameterReference $field)
+    protected FieldReference $field;
+
+    public function __construct(string | FieldReference $field)
     {
+        $this->field = $this->parseFieldReference($field);
     }
 
     public static function keyword(): string