+++ src/Support/TypeScriptTransformer/DataTypeScriptTransformer.php
@@ -17,6 +17,7 @@ use Spatie\LaravelData\Contracts\BaseData;
 use Spatie\LaravelData\Enums\DataTypeKind;
 use Spatie\LaravelData\Support\DataConfig;
 use Spatie\LaravelData\Support\DataProperty;
+use Spatie\LaravelData\Support\Lazy\ClosureLazy;
 use Spatie\LaravelTypeScriptTransformer\Transformers\DtoTransformer;
 use Spatie\TypeScriptTransformer\Attributes\Optional as TypeScriptOptional;
 use Spatie\TypeScriptTransformer\Structures\MissingSymbolsCollection;
@@ -70,7 +71,7 @@ class DataTypeScriptTransformer extends DtoTransformer
                     || $dataProperty->attributes->contains(
                         fn (object $attribute) => $attribute instanceof TypeScriptOptional
                     )
-                    || $dataProperty->type->isLazy
+                    || ($dataProperty->type->lazyType && $dataProperty->type->lazyType !== ClosureLazy::class)
                     || $dataProperty->type->isOptional;
 
                 $transformed = $this->typeToTypeScript(