+++ tests/DataTest.php
@@ -6,9 +6,12 @@ use Illuminate\Contracts\Support\Arrayable;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Foundation\Http\FormRequest;
 use Illuminate\Support\Collection;
+use Illuminate\Validation\ValidationException;
 use Inertia\LazyProp;
+use Spatie\LaravelData\Attributes\Computed;
 use Spatie\LaravelData\Attributes\DataCollectionOf;
 use Spatie\LaravelData\Attributes\MapOutputName;
+use Spatie\LaravelData\Attributes\Validation\Min;
 use Spatie\LaravelData\Attributes\WithCast;
 use Spatie\LaravelData\Attributes\WithCastable;
 use Spatie\LaravelData\Attributes\WithTransformer;
@@ -27,8 +30,10 @@ use Spatie\LaravelData\Contracts\DataObject;
 use Spatie\LaravelData\Data;
 use Spatie\LaravelData\DataCollection;
 use Spatie\LaravelData\Exceptions\CannotCreateData;
+use Spatie\LaravelData\Exceptions\CannotSetComputedValue;
 use Spatie\LaravelData\Lazy;
 use Spatie\LaravelData\Optional;
+use Spatie\LaravelData\Support\Lazy\ClosureLazy;
 use Spatie\LaravelData\Support\Lazy\InertiaLazy;
 use Spatie\LaravelData\Support\Transformation\TransformationContextFactory;
 use Spatie\LaravelData\Tests\Fakes\Castables\SimpleCastable;
@@ -332,6 +337,26 @@ it('always transforms lazy inertia data to inertia lazy props', function () {
     expect($data->toArray()['name'])->toBeInstanceOf(LazyProp::class);
 });
 
+it('always transforms closure lazy into closures for inertia', function () {
+    $blueprint = new class () extends Data {
+        public function __construct(
+            public string|ClosureLazy|null $name = null
+        ) {
+        }
+
+        public static function create(string $name): static
+        {
+            return new self(
+                Lazy::closure(fn () => $name)
+            );
+        }
+    };
+
+    $data = $blueprint::create('Freek');
+
+    expect($data->toArray()['name'])->toBeInstanceOf(Closure::class);
+});
+
 it('can get the empty version of a data object', function () {
     $dataClass = new class () extends Data {
         public string $property;
@@ -631,9 +656,11 @@ it('can append data via method overwriting with closures', function () {
 
         public function with(): array
         {
-            return ['alt_name' => static function (self $data) {
-                return $data->name . ' from Spatie via closure';
-            }];
+            return [
+                'alt_name' => static function (self $data) {
+                    return $data->name.' from Spatie via closure';
+                },
+            ];
         }
     };
 
@@ -651,9 +678,11 @@ test('when using additional method and with method additional method gets priori
 
         public function with(): array
         {
-            return ['alt_name' => static function (self $data) {
-                return $data->name . ' from Spatie via closure';
-            }];
+            return [
+                'alt_name' => static function (self $data) {
+                    return $data->name.' from Spatie via closure';
+                },
+            ];
         }
     };
 
@@ -2373,4 +2402,63 @@ it('can fetch non-lazy union data', function () {
     expect($data->simple->string)->toBe('A');
     expect($data->dataCollection->toCollection()->pluck('string')->toArray())->toBe(['B', 'C']);
     expect($data->fakeModel->string)->toBe('non-lazy');
+});
+
+it('can set a default value for data object', function () {
+    $dataObject = new class ('', '') extends Data {
+        #[Min(10)]
+        public string|Optional $full_name;
+
+        public function __construct(
+            public string $first_name,
+            public string $last_name,
+        ) {
+            $this->full_name = "{$this->first_name} {$this->last_name}";
+        }
+    };
+
+    expect($dataObject::from(['first_name' => 'Ruben', 'last_name' => 'Van Assche']))
+        ->first_name->toBe('Ruben')
+        ->last_name->toBe('Van Assche')
+        ->full_name->toBe('Ruben Van Assche');
+
+    expect($dataObject::from(['first_name' => 'Ruben', 'last_name' => 'Van Assche', 'full_name' => 'Ruben Versieck']))
+        ->first_name->toBe('Ruben')
+        ->last_name->toBe('Van Assche')
+        ->full_name->toBe('Ruben Versieck');
+
+    expect($dataObject::validateAndCreate(['first_name' => 'Ruben', 'last_name' => 'Van Assche']))
+        ->first_name->toBe('Ruben')
+        ->last_name->toBe('Van Assche')
+        ->full_name->toBe('Ruben Van Assche');
+
+    expect(fn () => $dataObject::validateAndCreate(['first_name' => 'Ruben', 'last_name' => 'Van Assche', 'full_name' => 'too short']))
+        ->toThrow(ValidationException::class);
+});
+
+it('can have a computed value', function () {
+    $dataObject = new class ('', '') extends Data {
+        #[Computed]
+        public string $full_name;
+
+        public function __construct(
+            public string $first_name,
+            public string $last_name,
+        ) {
+            $this->full_name = "{$this->first_name} {$this->last_name}";
+        }
+    };
+
+    expect($dataObject::from(['first_name' => 'Ruben', 'last_name' => 'Van Assche']))
+        ->first_name->toBe('Ruben')
+        ->last_name->toBe('Van Assche')
+        ->full_name->toBe('Ruben Van Assche');
+
+    expect($dataObject::validateAndCreate(['first_name' => 'Ruben', 'last_name' => 'Van Assche']))
+        ->first_name->toBe('Ruben')
+        ->last_name->toBe('Van Assche')
+        ->full_name->toBe('Ruben Van Assche');
+
+    expect(fn () => $dataObject::from(['first_name' => 'Ruben', 'last_name' => 'Van Assche', 'full_name' => 'Ruben Versieck']))
+        ->toThrow(CannotSetComputedValue::class);
 });